Use Vector3::new in first example
The `new` function on vectors is oddly hard to find in the documentation (the first time I tried using nalgebra I think I actually give up before finding it...). Given that creating a vector is likely one of the very first things beginners are likely to want to do, it is worth demonstrating it here.