Vector and quaternion alias improvements.
By aliasing to the aliases defined within glm rather than going
directly to nalgebra types / aliases, we're able to improve the
experience of linking through the docs in the glm docs.

For example, with TVec1 defined in terms of TVec rather than Vector1,
then the docs link from TVec1 to TVec and we're able to better
explain the concepts at the glm level.