Derive Copy, Eq, and Hash for Quaternion.
They were implemented manually.

Eq was implemented in such a way that -q was considered equal to q,
probably because they represent the same rotation. They are, however,
not equal quaternions. Because of this error, the implementation of Hash
was not compatible with the implementation of Eq. This fixes that.

Fixes #583.