Improve error message for `cargo test`
There is a custom compiler error, hinting that the "debug", "compare"
and "rand" feature are required for compiling and running the tests.
However, this error was not displayed when running `cargo test` due to
other compilation errors taking precedence.

This is now avoided by just not compiling the integration tests when the
necessary features are not enabled.