Refactor CsMatrix to support borrowed data / different index types
This combs through the CSR / CSC matrix formats, and unifies them into a
single type `CsMatrix`. While previously this type was private, this
refactor unifies them both into a single public generic type, that can
support either CSR or CSC formats with the same code.

The generic type introduces a few things that make a lot of code
redundant, or at least fairly verbose. Some of the key changes here:

- SparsityPattern was removed
- Iterators were slightly renamed / moved around (two new iterators were
introduced)
- CsrMatrix / CscMatrix are now aliases for the generic type
- Almost all functionality that would take `&mut self` of a CsXMatrix
has been removed.
- CsMatrix stores "offsets" as counts per lane, not accumulated offsets.
This may actually be wrong and in need of a future fixup, since it means
that calling `get_lane` will have different performance when doing
`get_lane(0)` vs. `get_lane(1000)`, as an example. However, this is on
the radar and is up for discussion, because it is also advantageous in
some scenarios (e.g. trying to slice along the major axis without
copying any data).

Let's go through each of these big changes one-by-one:

SparsityPattern
---------------

SparsityPattern was removed primarily because it adds nesting to the
generic type. Instead, we now take in `MajorOffsets` and `MinorIndices`
as generic parameters of the CsMatrix type. This is useful because
nesting the type in means:

1. We need to have interfaces for said type
2. The code for it needs to be supported
3. We have to understand the abstraction

In pretty much every case, `Borrow<[Offset]>` and `Borrow<[Index]>`
(which gives us `&[Offset]` and `&[Index]` respectively) solves the
first point. In the past these were just thin wrappers around vectors,
but since they are now made generic, we can get by without needing
SparsityPattern purely by using the slice interface.

2 is handled trivially (the Rust std library is what supports `Borrow`
and std::slice types), and 3 is something that you need to have a cogent
understanding of in order to get started with Rust (else face the borrow
checker). So overall, this just presents a cleaner way to offer the same
functionality while reducing the overall code footprint. It wasn't a
terrible abstraction, but nesting these types further into the
SparsityPattern struct makes a lot of the remaining code more burdensome
to write (because then we need to guarantee the first issue, an
interface, via some trait or generic impl).

Iterators
---------

Two new iterators were introduced:

1. AllElementsIter
2. CsLaneIter

The first one is fairly self-explanatory - it iterates through all
"entries" (including zeros) in the matrix. This isn't what you want most
of the time but is a direct analog to `get_entry`. Mostly, it's for the
sake of completion (otherwise you have to iterate through `get_entry`
which requires a lot of option unpacking).

The second one isn't "new" per se, since a type with that name existed
prior to the refactor. However, this refactor introduces iterators as:

- CsMatrixIter: iterate through each lane, as if calling `get_lane`
successively.
- CsLaneIter: iterate through all the (Index, Value) pairs in a given
lane.

The `CsLane` type has been removed because it was redundant. These
iterators allow us to implement `triplet_iter` in a much more concise
fashion. The new design along with these iterators allow for us to also
remove the `CsMatrix::filter` function, since it can be replaced by:

    mat.triplet_iter().filter(/* ... */);

Using idiomatic iterator code.

CsrMatrix / CscMatrix
---------------------

These two are now aliases / specializations of the CsMatrix type. This
is done to reduce nesting of the types, and is what enables the
`transpose` function to be generic to index / offset types as well,
since that function only needs to care about `shape` and
`CompressionKind`.

Not sure what else to say here, but the lack of nesting means we can get
rid of the source files csc.rs and csr.rs, since they mostly just
wrapped the CsMatrix type anyways.

The removal of `&mut self`
-------------------------

This one could be controversial. However, mutability in CsMatrix isn't
actually that useful. Hear me out, please. In any case where you have
mutability, you can really only have mutability on the non-zero values
in the matrix, not the entire matrix itself. This is fine if you're
adding a scalar to all non-zero values of the matrix, but doesn't make
sense in a lot of other contexts. Stuff like `cs_data_mut`,
`get_entry_mut`, etc. gave you a limited window into mutability when
mutable access to the underlying data isn't necessarily safe to have in
the compressed format.

Instead, it makes more sense to provide such functionality by consuming
the type (i.e. taking ownership OR using `self` instead of `&mut self`).
In that scenario, if the compressed layout needs to be recomputed
because something in the structure changes, then you as a consumer will
have `self` and can do that. You cannot change the underlying structure
if you have `&mut self`. I don't think this actually prevents anything
in the library, but rather just makes some of the clones / copies of
sparse types more explicit.

I think overall if there's a desire to maintain the ability to mutate
_just_ the non-zeros in the matrix, then a better way to do this would
be to make the CsMatrix type implement Copy, and to let that be
transparent where needed. However, I actually don't think we lose much
other than "mutable interfaces for the sake of a complete API," which
isn't necessarily semantically what you want.

Offsets as counts
-----------------

I did this because it was conceptually simpler, and because it avoids an
edge case in the new iterators (specifically, CsMatrixIter). The
iterator code is quite easy to understand IMO, and should be about as
fast as we can hope for. The one sad bit of this is that `get_lane` will
necessarily be slower when getting lanes at the end of the matrix rather
than lanes near the beginning.

However, in most cases I would say one should not be getting lanes using
`get_lane`. We can avoid this in CsMatrixIter by caching the previous
total offset and adding each lane offset as we go. This makes it
optimal, unlike `get_lane`. Pretty much every current operation will
want to iterate through each lane sequentially, so I think from a
computation standpoint it's fine if `get_lane` is slightly sub-optimal,
since the natural way to write code will be to call `CsMatrix::iter`
instead of `for i in 0..CsMatrix::nmajor` { CsMatrix::get_lane(i); } or
whatever.

Again, it's up for discussion, but it simplifies a vast majority of
lane-by-lane access while keeping it cache friendly :)