Add minor-lane iterators to CsMatrix type
This is primarily useful for implementing a generalized
sparse-matrix-multiply. Consider the following matrix multiplication:

    C := A <dot> B

In this matrix-matrix product, we can express the problem as iterating
through the rows of A, and taking the successive vector-vector
dot-product with the columns of B.

In such a scenario, a CSR <dot> CSC product is the most optimal sparse
matrix multiplication that we can perform, because the major lanes
aligne with what we want to iterate over. In such a scenario, we can
make the sparse matrix multiplication optimal (i.e. O(n^2 log(n)) in the
average case).

In other cases, we get slightly worse performance because we cannot
exploit the compressed structure of the type:

- CSC <dot> CSC: We must iterate across the minor lane of the first
matrix. This is slightly worse than the optimal case, but we can
actually do fine here since each row can be cached (i.e. we only need to
do an expensive re-ordering once).

- CSR <dot> CSR: This would imply that we need to iterate across the
minor lane of the right-most matrix. Since the right hand side cannot be
cached as in the CSC <dot> CSC case, this is significantly worse than
the optimal case. However, we can exploit the following property in
order to make this only as bad as the CSC <dot> CSC case:

    A <dot> B := (B' <dot> A')'

where the single quote represents a transpose. Since transposition is
free, we can convert all CSR <dot> CSR problems into CSC <dot> CSC
problems.

- CSC <dot> CSR: This is the worst case scenario for performance, since
it's exactly opposite to the structure we want to exploit. Because of
the compression being on exactly the wrong axes, we can't actually do
anything here to speed things up. Some things are cursed and always will
be. So for both matrices we have to iterate along the minor axis, which
is more expensive from a performance perspective than the optimal case.

Overall, these iterators make it possible to use (mostly) the same
algorithm for any of the above cases, expressed in terms of the
"fastest" possible matrix multiplication we can do given the underlying
structure of the types.

NOTE: Outside of spGEMM-like operations, these iterators should probably
be mostly avoided.