Relax Copy trait bound to Clone in spmm
It does the same thing, but theoretically opens the door to number types
that don't implement Copy for whatever reason. Still not sure if we
should tighten this trait bound to Scalar types for consistency. So far
I didn't because we don't actually need the Scalar guarantee and it's
not like it would change the performance since Rust doesn't allow for
adding references together anyways.