Remove Offset and Index generic args from CsMatrix
This would mostly be useful if you were trying to be compatible with a
C-library that was exposing the same format as `Borrow<[u16]>` or
`Borrow<[u32]>` or something. However, there is no advantage to this in
Rust when discussing actual native code, and the crate does not yet have
enough justification to allow for this functionality (in particular,
where is such and such C-library???).

So I'm just nixing it. You don't need generic index types here, and they
would mostly be a mistake because you have to then convert them up to
usize anyways once you do literally any arithmetic with a sparse matrix.
It's functionally useless -- you either upcast to usize or you have to
somehow guess at what operations are available on said type. One demand
was that these values could always be `Unsigned` per the num_traits
crate, but at what cost? Overflow / underflow aren't really guaranteed,
and there's no way to cleanly guarantee that code wouldn't panic.

NOTE: you also can't do that with usize, but we have the advantage with
usize that our vectors:

1. Index according to usize as a type
2. Can have usize::MAX number of elements contained therein

As a result, that informs us that offsets and indices are both
necessarily less than usize::MAX, because the data itself can only hold
usize::MAX values in this format. usize just makes sense.

There is definitely an argument that perhaps u32 is enough if you only
have e.g. small matrices, but then we need to clone + call `.into()` on
all these indices / offsets just to do very basic things. I reckon that
usize is not slow enough in benchmarks to justify this, but I could be
wrong.