Add cholesky factorization module back to crate
This re-implements the Choleskly factorization in terms of the new
`CsMatrix` type. This has a few advantages over the old version, but is
still WIP until we get the spsolve_lower_ stuff back in.

Firstly, the advantages:

1. This works for CSC and CSR matrix types now. Seriously, the matrix
has to be symmetric, so there's no reason that we should limit it to
either case. Since we know that this is only valid for `T: RealField`,
that means that we can make a lot of assumptions about the Cholesky
because the matrix is symmetric. If we want to consider complex numbers
in our matrices, then this won't work because we need to worry about
taking the conjugate of some of our values as well depending on the
orientation of the matrix. That was never supported here though, so we
don't worry about it.

2. We don't store several copies of the L / U / M sparsity patterns.
This is another benefit of getting rid of the sparsity pattern as a
type: you realize that all these patterns are similar (in the case of L
and U you discover they are the same), or can be better passed around
and used with the CsMatrix type. A "CholeskyPattern" struct is provided
to replace the "symbolic" Cholesky. This in total stops us from having
to store several sparsity patterns (memory improvement!!!) as well as
re-computing the transpose of the sparsity patterns in the L / U case. I
don't know why this was done initially -- the pattern is symmetric so
uh... it's the same pattern just re-computed in a very expensive manner.

Tests still need to be added (as they do everywhere) and the `solve` and
`solve_mut` bits still don't work. There is a question of how public we
want these functions to be. I think pub(crate) makes the most sense
for now, since otherwise we'd probably want to encapsulate them into two
traits for solve and solve_mut, and that would depend if the B matrix in
`A x = B` is sparse or dense. For now, we'll only cover the Dense-Matrix
case since that is the extent of current support.