Add unit tests for valid & invalid CsMatrix construction
These were taken from the current tests/ folder for the most part, but
adapted for the new type & format. I've left out a number of tests that
pertained to `_mut` functions, since those are no longer needed.

This file is now getting to the point where it really should be broken
into separate modules. The iterators could easily be in their own
submodule.