Add unit tests for COO <-> CS conversions
While porting these tests from the integration tests folder to the unit
tests folder, I realized that the current COO conversions were a bit
difficult to parse and had a lot going on. I simplified the process by
leveraging lexicographic sorting of tuples.

This removes a lot of code for handling duplicates / sorting lanes that
can otherwise be handled by standard library functions. I reckon that
this is faster because it requires less data partitioning and it becomes
an O(n * log(n) + n) in the worst case (assuming the worst possible
sorting).

Overall I think this is an improvement to both readability and
compactness of the code, but I've left some comments around the sort
detailing the lexicographic sorting nature of tuples.