Use matrixcompare functions instead of explicit comparisons
This makes the tests somewhat easier to read. We do rely on the
matrixcompare implementations in the matrixcompare submodule to be
working in order for this to be valid, but those are pretty standard and
easy to understand?

IDK, maybe there should be more tests there just to be sure, but I am
fairly certain that it is working as expected.