Re-enable integration tests that were not moved over
This is mostly just the tests for COO matrices, which were untouched in
this refactor. Future work should port these to unit tests.

Additionally, there's some code overlap between whats in those tests and
what's now in the proptest submodule of the crate. However, I left that
alone for now in the interest of getting this work closed out.