fix: Add reproducible_smatrix()
Some algorithms may not converge when used on completely random values
with the default value of epsilon and unlimited iterations.

`reproducible_dmatrix()` already exist to circumvent this for `DMatrix`,
so I implemented the same for `SMatrix`.

In my tests this problem manifested itself only on
`schur_decompose_4x4`, but I decided to apply similar fix for all
benchmarks that also use `reproducible_dmatrix()` for `DMatrix`.